Possible fix for #1148
