+++ mm2src/docker_tests.rs
@@ -43,6 +43,10 @@ fn main() { unimplemented!() }
 /// rustfmt cannot resolve the module path within docker_tests.
 /// Specify the path manually outside the docker_tests.
 #[cfg(rustfmt)]
+#[path = "docker_tests/docker_ordermatch_tests.rs"]
+mod docker_ordermatch_tests;
+
+#[cfg(rustfmt)]
 #[path = "docker_tests/swaps_confs_settings_sync_tests.rs"]
 mod swaps_confs_settings_sync_tests;
 
@@ -80,6 +84,8 @@ mod docker_tests {
 
 #[cfg(all(test, not(target_arch = "wasm32")))]
 mod docker_tests {
+    #[rustfmt::skip]
+    mod docker_ordermatch_tests;
     #[rustfmt::skip]
     mod qrc20_tests;
     #[rustfmt::skip]
@@ -3811,617 +3817,6 @@ mod docker_tests {
 
         block_on(mm_bob.wait_for_log(22., |log| log.contains("Entering the maker_swap_loop MYCOIN/ETH"))).unwrap();
         block_on(mm_alice.wait_for_log(22., |log| log.contains("Entering the taker_swap_loop MYCOIN/ETH"))).unwrap();
-
-        block_on(mm_bob.stop()).unwrap();
-        block_on(mm_alice.stop()).unwrap();
-    }
-
-    fn check_asks_num(mm: &MarketMakerIt, base: &str, rel: &str, expected: usize) {
-        log!({"Get {}/{} orderbook", base, rel});
-        let rc = block_on(mm.rpc(json! ({
-            "userpass": mm.userpass,
-            "method": "orderbook",
-            "base": base,
-            "rel": rel,
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-        let orderbook: OrderbookResponse = json::from_str(&rc.1).unwrap();
-        log!("orderbook "[orderbook]);
-        assert_eq!(
-            orderbook.asks.len(),
-            expected,
-            "{}/{} orderbook must have exactly {} ask(s)",
-            base,
-            rel,
-            expected
-        );
-    }
-
-    fn check_bids_num(mm: &MarketMakerIt, base: &str, rel: &str, expected: usize) {
-        log!({"Get {}/{} orderbook", base, rel});
-        let rc = block_on(mm.rpc(json! ({
-            "userpass": mm.userpass,
-            "method": "orderbook",
-            "base": base,
-            "rel": rel,
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
-        let orderbook: OrderbookResponse = json::from_str(&rc.1).unwrap();
-        log!("orderbook "[orderbook]);
-        assert_eq!(
-            orderbook.bids.len(),
-            expected,
-            "{}/{} orderbook must have exactly {} bid(s)",
-            base,
-            rel,
-            expected
-        );
-    }
-
-    fn check_orderbook_depth(mm: &MarketMakerIt, pairs: &[(&str, &str)], expected: &[(usize, usize)]) {
-        log!({"Get {:?} orderbook depth", pairs});
-        let rc = block_on(mm.rpc(json! ({
-            "userpass": mm.userpass,
-            "method": "orderbook_depth",
-            "pairs": pairs,
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!orderbook_depth: {}", rc.1);
-        let orderbook_depth: OrderbookDepthResponse = json::from_str(&rc.1).unwrap();
-        log!("orderbook depth "[orderbook_depth]);
-        for (pair, expected_depth) in pairs.iter().zip(expected) {
-            let actual_depth = orderbook_depth
-                .result
-                .iter()
-                .find(|pair_with_depth| pair.0 == pair_with_depth.pair.0 && pair.1 == pair_with_depth.pair.1)
-                .unwrap_or_else(|| panic!("Orderbook depth result doesn't contain pair {:?}", pair));
-
-            assert_eq!(
-                actual_depth.depth.asks, expected_depth.0,
-                "expected {} asks for pair {:?}",
-                expected_depth.0, actual_depth.pair,
-            );
-            assert_eq!(
-                actual_depth.depth.bids, expected_depth.1,
-                "expected {} bids for pair {:?}",
-                expected_depth.1, actual_depth.pair,
-            );
-        }
-    }
-
-    fn check_best_orders(
-        mm: &MarketMakerIt,
-        action: &str,
-        for_coin: &str,
-        ticker_in_response: &str,
-        expected_num_orders: usize,
-    ) {
-        log!({"Get best orders for {}", for_coin});
-        let rc = block_on(mm.rpc(json! ({
-            "userpass": mm.userpass,
-            "method": "best_orders",
-            "coin": for_coin,
-            "action": action,
-            "volume": 1,
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!best_orders: {}", rc.1);
-        let best_orders: BestOrdersResponse = json::from_str(&rc.1).unwrap();
-        let orders = best_orders
-            .result
-            .get(ticker_in_response)
-            .unwrap_or_else(|| panic!("No orders for ticker {}", ticker_in_response));
-
-        assert_eq!(orders.len(), expected_num_orders);
-    }
-
-    #[test]
-    fn test_ordermatch_custom_orderbook_ticker_both_on_maker() {
-        let (_ctx, _, bob_priv_key) = generate_coin_with_random_privkey("MYCOIN", 1000.into());
-        let (_ctx, _, alice_priv_key) = generate_coin_with_random_privkey("MYCOIN1", 2000.into());
-        let coins = json! ([
-            {"coin":"MYCOIN", "asset":"MYCOIN","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-            {"coin":"MYCOIN1", "asset":"MYCOIN1","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-            {"coin":"MYCOIN-Custom", "orderbook_ticker":"MYCOIN", "asset":"MYCOIN","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-            {"coin":"MYCOIN1-Custom", "orderbook_ticker":"MYCOIN1", "asset":"MYCOIN1","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-        ]);
-        let mut mm_bob = MarketMakerIt::start(
-            json! ({
-                "gui": "nogui",
-                "netid": 9000,
-                "dht": "on",  // Enable DHT without delay.
-                "passphrase": format!("0x{}", hex::encode(bob_priv_key)),
-                "coins": coins,
-                "rpc_password": "pass",
-                "i_am_seed": true,
-            }),
-            "pass".to_string(),
-            None,
-        )
-        .unwrap();
-        let (_bob_dump_log, _bob_dump_dashboard) = mm_dump(&mm_bob.log_path);
-
-        let mut mm_alice = MarketMakerIt::start(
-            json! ({
-                "gui": "nogui",
-                "netid": 9000,
-                "dht": "on",  // Enable DHT without delay.
-                "passphrase": format!("0x{}", hex::encode(alice_priv_key)),
-                "coins": coins,
-                "rpc_password": "pass",
-                "seednodes": vec![format!("{}", mm_bob.ip)],
-            }),
-            "pass".to_string(),
-            None,
-        )
-        .unwrap();
-        let (_alice_dump_log, _alice_dump_dashboard) = mm_dump(&mm_alice.log_path);
-
-        log!([block_on(enable_native(&mm_bob, "MYCOIN-Custom", &[]))]);
-        log!([block_on(enable_native(&mm_bob, "MYCOIN1-Custom", &[]))]);
-        log!([block_on(enable_native(&mm_alice, "MYCOIN", &[]))]);
-        log!([block_on(enable_native(&mm_alice, "MYCOIN1", &[]))]);
-        let rc = block_on(mm_bob.rpc(json! ({
-            "userpass": mm_bob.userpass,
-            "method": "setprice",
-            "base": "MYCOIN-Custom",
-            "rel": "MYCOIN1-Custom",
-            "price": 1,
-            "max": true,
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-        let sell_result: SetPriceResponse = json::from_str(&rc.1).unwrap();
-        assert_eq!(sell_result.result.base, "MYCOIN-Custom");
-        assert_eq!(sell_result.result.rel, "MYCOIN1-Custom");
-        assert_eq!(sell_result.result.base_orderbook_ticker, Some("MYCOIN".to_owned()));
-        assert_eq!(sell_result.result.rel_orderbook_ticker, Some("MYCOIN1".to_owned()));
-
-        // check bob orders
-        check_asks_num(&mm_bob, "MYCOIN-Custom", "MYCOIN1-Custom", 1);
-        check_asks_num(&mm_bob, "MYCOIN", "MYCOIN1", 1);
-        check_bids_num(&mm_bob, "MYCOIN1-Custom", "MYCOIN-Custom", 1);
-        check_bids_num(&mm_bob, "MYCOIN1", "MYCOIN", 1);
-
-        // check multiple RPCs on alice side
-        check_best_orders(&mm_alice, "sell", "MYCOIN1", "MYCOIN", 1);
-        check_best_orders(&mm_alice, "sell", "MYCOIN1-Custom", "MYCOIN", 1);
-        check_best_orders(&mm_alice, "buy", "MYCOIN", "MYCOIN1", 1);
-        check_best_orders(&mm_alice, "buy", "MYCOIN-Custom", "MYCOIN1", 1);
-
-        check_orderbook_depth(
-            &mm_alice,
-            &[
-                ("MYCOIN", "MYCOIN1"),
-                ("MYCOIN", "MYCOIN1-Custom"),
-                ("MYCOIN-Custom", "MYCOIN1"),
-                ("MYCOIN-Custom", "MYCOIN1-Custom"),
-            ],
-            &[(1, 0); 4],
-        );
-
-        check_asks_num(&mm_alice, "MYCOIN-Custom", "MYCOIN1-Custom", 1);
-        check_asks_num(&mm_alice, "MYCOIN", "MYCOIN1", 1);
-        check_bids_num(&mm_alice, "MYCOIN1-Custom", "MYCOIN-Custom", 1);
-        check_bids_num(&mm_alice, "MYCOIN1", "MYCOIN", 1);
-
-        // check orderbook depth again after subscription to the topic
-        check_orderbook_depth(
-            &mm_alice,
-            &[
-                ("MYCOIN", "MYCOIN1"),
-                ("MYCOIN", "MYCOIN1-Custom"),
-                ("MYCOIN-Custom", "MYCOIN1"),
-                ("MYCOIN-Custom", "MYCOIN1-Custom"),
-            ],
-            &[(1, 0); 4],
-        );
-
-        let rc = block_on(mm_alice.rpc(json! ({
-            "userpass": mm_alice.userpass,
-            "method": "buy",
-            "base": "MYCOIN",
-            "rel": "MYCOIN1",
-            "price": 1,
-            "volume": "999.99999",
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!buy: {}", rc.1);
-
-        block_on(mm_bob.wait_for_log(22., |log| {
-            log.contains("Entering the maker_swap_loop MYCOIN-Custom/MYCOIN1-Custom")
-        }))
-        .unwrap();
-        block_on(mm_alice.wait_for_log(22., |log| log.contains("Entering the taker_swap_loop MYCOIN/MYCOIN1")))
-            .unwrap();
-
-        block_on(mm_bob.stop()).unwrap();
-        block_on(mm_alice.stop()).unwrap();
-    }
-
-    #[test]
-    fn test_ordermatch_custom_orderbook_ticker_both_on_taker() {
-        let (_ctx, _, bob_priv_key) = generate_coin_with_random_privkey("MYCOIN", 1000.into());
-        let (_ctx, _, alice_priv_key) = generate_coin_with_random_privkey("MYCOIN1", 2000.into());
-
-        let coins = json! ([
-            {"coin":"MYCOIN", "asset":"MYCOIN","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-            {"coin":"MYCOIN1", "asset":"MYCOIN1","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-            {"coin":"MYCOIN-Custom", "orderbook_ticker":"MYCOIN", "asset":"MYCOIN","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-            {"coin":"MYCOIN1-Custom", "orderbook_ticker":"MYCOIN1", "asset":"MYCOIN1","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-        ]);
-        let mut mm_bob = MarketMakerIt::start(
-            json! ({
-                "gui": "nogui",
-                "netid": 9000,
-                "dht": "on",  // Enable DHT without delay.
-                "passphrase": format!("0x{}", hex::encode(bob_priv_key)),
-                "coins": coins,
-                "rpc_password": "pass",
-                "i_am_seed": true,
-            }),
-            "pass".to_string(),
-            None,
-        )
-        .unwrap();
-        let (_bob_dump_log, _bob_dump_dashboard) = mm_dump(&mm_bob.log_path);
-
-        let mut mm_alice = MarketMakerIt::start(
-            json! ({
-                "gui": "nogui",
-                "netid": 9000,
-                "dht": "on",  // Enable DHT without delay.
-                "passphrase": format!("0x{}", hex::encode(alice_priv_key)),
-                "coins": coins,
-                "rpc_password": "pass",
-                "seednodes": vec![format!("{}", mm_bob.ip)],
-            }),
-            "pass".to_string(),
-            None,
-        )
-        .unwrap();
-        let (_alice_dump_log, _alice_dump_dashboard) = mm_dump(&mm_alice.log_path);
-
-        log!([block_on(enable_native(&mm_bob, "MYCOIN", &[]))]);
-        log!([block_on(enable_native(&mm_bob, "MYCOIN1", &[]))]);
-        log!([block_on(enable_native(&mm_alice, "MYCOIN-Custom", &[]))]);
-        log!([block_on(enable_native(&mm_alice, "MYCOIN1-Custom", &[]))]);
-        let rc = block_on(mm_bob.rpc(json! ({
-            "userpass": mm_bob.userpass,
-            "method": "setprice",
-            "base": "MYCOIN",
-            "rel": "MYCOIN1",
-            "price": 1,
-            "max": true,
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-
-        // check orderbooks on bob side
-        check_asks_num(&mm_bob, "MYCOIN-Custom", "MYCOIN1-Custom", 1);
-        check_asks_num(&mm_bob, "MYCOIN", "MYCOIN1", 1);
-        check_bids_num(&mm_bob, "MYCOIN1-Custom", "MYCOIN-Custom", 1);
-        check_bids_num(&mm_bob, "MYCOIN1", "MYCOIN", 1);
-
-        // check multiple RPCs on alice side
-        check_best_orders(&mm_alice, "sell", "MYCOIN1", "MYCOIN", 1);
-        check_best_orders(&mm_alice, "sell", "MYCOIN1-Custom", "MYCOIN", 1);
-        check_best_orders(&mm_alice, "buy", "MYCOIN", "MYCOIN1", 1);
-        check_best_orders(&mm_alice, "buy", "MYCOIN-Custom", "MYCOIN1", 1);
-
-        check_orderbook_depth(
-            &mm_alice,
-            &[
-                ("MYCOIN", "MYCOIN1"),
-                ("MYCOIN", "MYCOIN1-Custom"),
-                ("MYCOIN-Custom", "MYCOIN1"),
-                ("MYCOIN-Custom", "MYCOIN1-Custom"),
-            ],
-            &[(1, 0); 4],
-        );
-
-        check_asks_num(&mm_alice, "MYCOIN-Custom", "MYCOIN1-Custom", 1);
-        check_asks_num(&mm_alice, "MYCOIN", "MYCOIN1", 1);
-        check_bids_num(&mm_alice, "MYCOIN1-Custom", "MYCOIN-Custom", 1);
-        check_bids_num(&mm_alice, "MYCOIN1", "MYCOIN", 1);
-
-        check_orderbook_depth(
-            &mm_alice,
-            &[
-                ("MYCOIN", "MYCOIN1"),
-                ("MYCOIN", "MYCOIN1-Custom"),
-                ("MYCOIN-Custom", "MYCOIN1"),
-                ("MYCOIN-Custom", "MYCOIN1-Custom"),
-            ],
-            &[(1, 0); 4],
-        );
-
-        let rc = block_on(mm_alice.rpc(json! ({
-            "userpass": mm_alice.userpass,
-            "method": "buy",
-            "base": "MYCOIN-Custom",
-            "rel": "MYCOIN1-Custom",
-            "price": 1,
-            "volume": "999.99999",
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!buy: {}", rc.1);
-        let buy_result: BuyOrSellRpcResult = json::from_str(&rc.1).unwrap();
-        assert_eq!(buy_result.result.base, "MYCOIN-Custom");
-        assert_eq!(buy_result.result.rel, "MYCOIN1-Custom");
-        assert_eq!(buy_result.result.base_orderbook_ticker, Some("MYCOIN".to_owned()));
-        assert_eq!(buy_result.result.rel_orderbook_ticker, Some("MYCOIN1".to_owned()));
-
-        block_on(mm_bob.wait_for_log(22., |log| log.contains("Entering the maker_swap_loop MYCOIN/MYCOIN1"))).unwrap();
-        block_on(mm_alice.wait_for_log(22., |log| {
-            log.contains("Entering the taker_swap_loop MYCOIN-Custom/MYCOIN1-Custom")
-        }))
-        .unwrap();
-
-        block_on(mm_bob.stop()).unwrap();
-        block_on(mm_alice.stop()).unwrap();
-    }
-
-    #[test]
-    fn test_ordermatch_custom_orderbook_ticker_mixed_case_one() {
-        let (_ctx, _, bob_priv_key) = generate_coin_with_random_privkey("MYCOIN", 1000.into());
-        let (_ctx, _, alice_priv_key) = generate_coin_with_random_privkey("MYCOIN1", 2000.into());
-
-        let coins = json! ([
-            {"coin":"MYCOIN", "asset":"MYCOIN","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-            {"coin":"MYCOIN1", "asset":"MYCOIN1","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-            {"coin":"MYCOIN-Custom", "orderbook_ticker":"MYCOIN", "asset":"MYCOIN","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-            {"coin":"MYCOIN1-Custom", "orderbook_ticker":"MYCOIN1", "asset":"MYCOIN1","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-        ]);
-        let mut mm_bob = MarketMakerIt::start(
-            json! ({
-                "gui": "nogui",
-                "netid": 9000,
-                "dht": "on",  // Enable DHT without delay.
-                "passphrase": format!("0x{}", hex::encode(bob_priv_key)),
-                "coins": coins,
-                "rpc_password": "pass",
-                "i_am_seed": true,
-            }),
-            "pass".to_string(),
-            None,
-        )
-        .unwrap();
-        let (_bob_dump_log, _bob_dump_dashboard) = mm_dump(&mm_bob.log_path);
-
-        let mut mm_alice = MarketMakerIt::start(
-            json! ({
-                "gui": "nogui",
-                "netid": 9000,
-                "dht": "on",  // Enable DHT without delay.
-                "passphrase": format!("0x{}", hex::encode(alice_priv_key)),
-                "coins": coins,
-                "rpc_password": "pass",
-                "seednodes": vec![format!("{}", mm_bob.ip)],
-            }),
-            "pass".to_string(),
-            None,
-        )
-        .unwrap();
-        let (_alice_dump_log, _alice_dump_dashboard) = mm_dump(&mm_alice.log_path);
-
-        log!([block_on(enable_native(&mm_bob, "MYCOIN-Custom", &[]))]);
-        log!([block_on(enable_native(&mm_bob, "MYCOIN1", &[]))]);
-        log!([block_on(enable_native(&mm_alice, "MYCOIN", &[]))]);
-        log!([block_on(enable_native(&mm_alice, "MYCOIN1-Custom", &[]))]);
-        let rc = block_on(mm_bob.rpc(json! ({
-            "userpass": mm_bob.userpass,
-            "method": "setprice",
-            "base": "MYCOIN-Custom",
-            "rel": "MYCOIN1",
-            "price": 1,
-            "max": true,
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-        let set_price: SetPriceResponse = json::from_str(&rc.1).unwrap();
-        assert_eq!(set_price.result.base, "MYCOIN-Custom");
-        assert_eq!(set_price.result.rel, "MYCOIN1");
-        assert_eq!(set_price.result.base_orderbook_ticker, Some("MYCOIN".to_owned()));
-        assert!(set_price.result.rel_orderbook_ticker.is_none());
-
-        // check orderbooks on bob side
-        check_asks_num(&mm_bob, "MYCOIN-Custom", "MYCOIN1-Custom", 1);
-        check_asks_num(&mm_bob, "MYCOIN", "MYCOIN1", 1);
-        check_bids_num(&mm_bob, "MYCOIN1-Custom", "MYCOIN-Custom", 1);
-        check_bids_num(&mm_bob, "MYCOIN1", "MYCOIN", 1);
-
-        // check multiple RPCs on alice side
-        check_best_orders(&mm_alice, "sell", "MYCOIN1", "MYCOIN", 1);
-        check_best_orders(&mm_alice, "sell", "MYCOIN1-Custom", "MYCOIN", 1);
-        check_best_orders(&mm_alice, "buy", "MYCOIN", "MYCOIN1", 1);
-        check_best_orders(&mm_alice, "buy", "MYCOIN-Custom", "MYCOIN1", 1);
-
-        check_orderbook_depth(
-            &mm_alice,
-            &[
-                ("MYCOIN", "MYCOIN1"),
-                ("MYCOIN", "MYCOIN1-Custom"),
-                ("MYCOIN-Custom", "MYCOIN1"),
-                ("MYCOIN-Custom", "MYCOIN1-Custom"),
-            ],
-            &[(1, 0); 4],
-        );
-
-        check_asks_num(&mm_alice, "MYCOIN-Custom", "MYCOIN1-Custom", 1);
-        check_asks_num(&mm_alice, "MYCOIN", "MYCOIN1", 1);
-        check_bids_num(&mm_alice, "MYCOIN1-Custom", "MYCOIN-Custom", 1);
-        check_bids_num(&mm_alice, "MYCOIN1", "MYCOIN", 1);
-
-        check_orderbook_depth(
-            &mm_alice,
-            &[
-                ("MYCOIN", "MYCOIN1"),
-                ("MYCOIN", "MYCOIN1-Custom"),
-                ("MYCOIN-Custom", "MYCOIN1"),
-                ("MYCOIN-Custom", "MYCOIN1-Custom"),
-            ],
-            &[(1, 0); 4],
-        );
-
-        let rc = block_on(mm_alice.rpc(json! ({
-            "userpass": mm_alice.userpass,
-            "method": "buy",
-            "base": "MYCOIN",
-            "rel": "MYCOIN1-Custom",
-            "price": 1,
-            "volume": "999.99999",
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!buy: {}", rc.1);
-        let buy_result: BuyOrSellRpcResult = json::from_str(&rc.1).unwrap();
-        assert_eq!(buy_result.result.base, "MYCOIN");
-        assert_eq!(buy_result.result.rel, "MYCOIN1-Custom");
-        assert!(buy_result.result.base_orderbook_ticker.is_none());
-        assert_eq!(buy_result.result.rel_orderbook_ticker, Some("MYCOIN1".to_owned()));
-
-        block_on(mm_bob.wait_for_log(22., |log| {
-            log.contains("Entering the maker_swap_loop MYCOIN-Custom/MYCOIN1")
-        }))
-        .unwrap();
-        block_on(mm_alice.wait_for_log(22., |log| {
-            log.contains("Entering the taker_swap_loop MYCOIN/MYCOIN1-Custom")
-        }))
-        .unwrap();
-
-        block_on(mm_bob.stop()).unwrap();
-        block_on(mm_alice.stop()).unwrap();
-    }
-
-    #[test]
-    fn test_ordermatch_custom_orderbook_ticker_mixed_case_two() {
-        let (_ctx, _, bob_priv_key) = generate_coin_with_random_privkey("MYCOIN", 1000.into());
-        let (_ctx, _, alice_priv_key) = generate_coin_with_random_privkey("MYCOIN1", 2000.into());
-
-        let coins = json! ([
-            {"coin":"MYCOIN", "asset":"MYCOIN","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-            {"coin":"MYCOIN1", "asset":"MYCOIN1","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-            {"coin":"MYCOIN-Custom", "orderbook_ticker":"MYCOIN", "asset":"MYCOIN","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-            {"coin":"MYCOIN1-Custom", "orderbook_ticker":"MYCOIN1", "asset":"MYCOIN1","txversion":4,"overwintered":1,"txfee":1000,"protocol":{"type":"UTXO"}},
-        ]);
-        let mut mm_bob = MarketMakerIt::start(
-            json! ({
-                "gui": "nogui",
-                "netid": 9000,
-                "dht": "on",  // Enable DHT without delay.
-                "passphrase": format!("0x{}", hex::encode(bob_priv_key)),
-                "coins": coins,
-                "rpc_password": "pass",
-                "i_am_seed": true,
-            }),
-            "pass".to_string(),
-            None,
-        )
-        .unwrap();
-        let (_bob_dump_log, _bob_dump_dashboard) = mm_dump(&mm_bob.log_path);
-
-        let mut mm_alice = MarketMakerIt::start(
-            json! ({
-                "gui": "nogui",
-                "netid": 9000,
-                "dht": "on",  // Enable DHT without delay.
-                "passphrase": format!("0x{}", hex::encode(alice_priv_key)),
-                "coins": coins,
-                "rpc_password": "pass",
-                "seednodes": vec![format!("{}", mm_bob.ip)],
-            }),
-            "pass".to_string(),
-            None,
-        )
-        .unwrap();
-        let (_alice_dump_log, _alice_dump_dashboard) = mm_dump(&mm_alice.log_path);
-
-        log!([block_on(enable_native(&mm_bob, "MYCOIN", &[]))]);
-        log!([block_on(enable_native(&mm_bob, "MYCOIN1-Custom", &[]))]);
-        log!([block_on(enable_native(&mm_alice, "MYCOIN-Custom", &[]))]);
-        log!([block_on(enable_native(&mm_alice, "MYCOIN1", &[]))]);
-        let rc = block_on(mm_bob.rpc(json! ({
-            "userpass": mm_bob.userpass,
-            "method": "setprice",
-            "base": "MYCOIN",
-            "rel": "MYCOIN1-Custom",
-            "price": 1,
-            "max": true,
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!setprice: {}", rc.1);
-        let set_price: SetPriceResponse = json::from_str(&rc.1).unwrap();
-        assert_eq!(set_price.result.base, "MYCOIN");
-        assert_eq!(set_price.result.rel, "MYCOIN1-Custom");
-        assert!(set_price.result.base_orderbook_ticker.is_none());
-        assert_eq!(set_price.result.rel_orderbook_ticker, Some("MYCOIN1".to_owned()));
-
-        // check orderbooks on bob side
-        check_asks_num(&mm_bob, "MYCOIN-Custom", "MYCOIN1-Custom", 1);
-        check_asks_num(&mm_bob, "MYCOIN", "MYCOIN1", 1);
-        check_bids_num(&mm_bob, "MYCOIN1-Custom", "MYCOIN-Custom", 1);
-        check_bids_num(&mm_bob, "MYCOIN1", "MYCOIN", 1);
-
-        // check multiple RPCs on alice side
-        check_best_orders(&mm_alice, "sell", "MYCOIN1", "MYCOIN", 1);
-        check_best_orders(&mm_alice, "sell", "MYCOIN1-Custom", "MYCOIN", 1);
-        check_best_orders(&mm_alice, "buy", "MYCOIN", "MYCOIN1", 1);
-        check_best_orders(&mm_alice, "buy", "MYCOIN-Custom", "MYCOIN1", 1);
-
-        check_orderbook_depth(
-            &mm_alice,
-            &[
-                ("MYCOIN", "MYCOIN1"),
-                ("MYCOIN", "MYCOIN1-Custom"),
-                ("MYCOIN-Custom", "MYCOIN1"),
-                ("MYCOIN-Custom", "MYCOIN1-Custom"),
-            ],
-            &[(1, 0); 4],
-        );
-
-        check_asks_num(&mm_alice, "MYCOIN-Custom", "MYCOIN1-Custom", 1);
-        check_asks_num(&mm_alice, "MYCOIN", "MYCOIN1", 1);
-        check_bids_num(&mm_alice, "MYCOIN1-Custom", "MYCOIN-Custom", 1);
-        check_bids_num(&mm_alice, "MYCOIN1", "MYCOIN", 1);
-
-        check_orderbook_depth(
-            &mm_alice,
-            &[
-                ("MYCOIN", "MYCOIN1"),
-                ("MYCOIN", "MYCOIN1-Custom"),
-                ("MYCOIN-Custom", "MYCOIN1"),
-                ("MYCOIN-Custom", "MYCOIN1-Custom"),
-            ],
-            &[(1, 0); 4],
-        );
-
-        let rc = block_on(mm_alice.rpc(json! ({
-            "userpass": mm_alice.userpass,
-            "method": "buy",
-            "base": "MYCOIN-Custom",
-            "rel": "MYCOIN1",
-            "price": 1,
-            "volume": "999.99999",
-        })))
-        .unwrap();
-        assert!(rc.0.is_success(), "!buy: {}", rc.1);
-        let buy_result: BuyOrSellRpcResult = json::from_str(&rc.1).unwrap();
-        assert_eq!(buy_result.result.base, "MYCOIN-Custom");
-        assert_eq!(buy_result.result.rel, "MYCOIN1");
-        assert_eq!(buy_result.result.base_orderbook_ticker, Some("MYCOIN".to_owned()));
-        assert!(buy_result.result.rel_orderbook_ticker.is_none());
-
-        block_on(mm_bob.wait_for_log(22., |log| {
-            log.contains("Entering the maker_swap_loop MYCOIN/MYCOIN1-Custom")
-        }))
-        .unwrap();
-        block_on(mm_alice.wait_for_log(22., |log| {
-            log.contains("Entering the taker_swap_loop MYCOIN-Custom/MYCOIN1")
-        }))
-        .unwrap();
 
         block_on(mm_bob.stop()).unwrap();
         block_on(mm_alice.stop()).unwrap();
